feat: Location Page & My Page UI improvements
Maintained the feature that displays the address only up to the district level.

Improved alignment and styling of UI elements.

Adjusted the layout of the current location section to enhance user experience.

Unified the styling of the current location section in the My Page for consistency.

Applied responsive design for better usability across devices.