Feat: Display chat room creation date in HeaderBar
This commit introduces the display of the chat room creation date in the HeaderBar.
A `formatDate` function was added to format the date string into "YYYY년 MM월 DD일" format.
The HeaderBar title now dynamically displays the formatted creation date of the current chat room, or "새로운 대화" for new chats.
Also, the left button prop was changed to leftCustomButton to allow custom buttons.